move bygning aggregation to separate extension file and create a basic test for it
